pyproject.toml
[tool.poetry]
name = "pdf-bot"
version = "0.1.0"
description = "An AI assistant for PDF files"
authors = ["ivan-pua-tpg <ivan.pua@tpgtelecom.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.2.0"
langchain-voyageai = "^0.1.1"
pypdf = "^4.3.1"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0.post1"
transformers = "^4.45.2"
torch = "^2.4.1"
nomkl = "^0.0.3"
langchain-huggingface = "^0.0.1"
huggingface-hub = {extras = ["cli"], version = "^0.25.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
